+++ tuple.go
@@ -141,7 +141,7 @@ func SequenceFieldEncode(field reflect.Value, buffer *bytes.Buffer) {
 	// TODO: Sequence[Dictionary[T1, T2]]
 	case reflect.TypeOf(*new(Sequence[Bool])):
 		size := field.Len()
-		ToCompact(uint64(size)).Encode(buffer)
+		ToCompact(size).Encode(buffer)
 		for i := 0; i < field.Len(); i++ {
 			SequenceFieldEncode(field.Index(i), buffer)
 		}
@@ -157,7 +157,7 @@ func SequenceFieldEncode(field reflect.Value, buffer *bytes.Buffer) {
 			// since there are infinite number of T we can't use switch
 			size := field.Len()
 			// TODO: if it is a FixedSequence[T] don't encode the length
-			ToCompact(uint64(size)).Encode(buffer)
+			ToCompact(size).Encode(buffer)
 			for i := 0; i < size; i++ {
 				seqElem := field.Index(i)
 				EncodeTuple(seqElem.Interface(), buffer)
@@ -211,7 +211,7 @@ func DictionaryFieldEncode(field reflect.Value, buffer *bytes.Buffer) {
 
 			// TODO: if it is a FixedSequence[T] don't encode the length
 			size := field.Len()
-			ToCompact(uint64(size)).Encode(buffer)
+			ToCompact(size).Encode(buffer)
 			for i := 0; i < size; i++ {
 				seqElem := field.Index(i)
 				EncodeTuple(seqElem.Interface(), buffer)
